# apps/frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workspaces/{wsId}/test-plans": {
      "get": {
        "operationId": "TestPlansController_list",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List test plans (optional ?releaseId=)",
        "tags": [
          "Test Plans"
        ]
      },
      "post": {
        "operationId": "TestPlansController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a test plan (named test batch)",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}": {
      "get": {
        "operationId": "TestPlansController_get",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get test plan detail (cases + progress summary)",
        "tags": [
          "Test Plans"
        ]
      },
      "patch": {
        "operationId": "TestPlansController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update test plan",
        "tags": [
          "Test Plans"
        ]
      },
      "delete": {
        "operationId": "TestPlansController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete test plan",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}/status": {
      "patch": {
        "operationId": "TestPlansController_updateStatus",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestPlanStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update plan status (DRAFT/ACTIVE/COMPLETED/ARCHIVED)",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}/cases": {
      "post": {
        "operationId": "TestPlansController_addCases",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlanCasesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add cases to plan (by ids or batch by releaseId)",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}/cases/{testCaseId}": {
      "delete": {
        "operationId": "TestPlansController_removeCase",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a case from plan",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}/start-run": {
      "post": {
        "operationId": "TestPlansController_startRun",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a walkthrough run from plan (reset cases to UNTESTED for a release)",
        "tags": [
          "Test Plans"
        ]
      }
    },
    "/api/workspaces/{wsId}/test-plans/{id}/walkthrough": {
      "get": {
        "operationId": "TestPlansController_walkthrough",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Walkthrough data: cases with steps + run status (prev/next navigation)",
        "tags": [
          "Test Plans"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/themes": {
      "get": {
        "operationId": "ThemesController_findAll",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List themes with linked entity count and aggregated votes",
        "tags": [
          "Themes"
        ]
      },
      "post": {
        "operationId": "ThemesController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThemeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a theme",
        "tags": [
          "Themes"
        ]
      }
    },
    "/api/workspaces/{wsId}/themes/{id}": {
      "patch": {
        "operationId": "ThemesController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a theme",
        "tags": [
          "Themes"
        ]
      },
      "delete": {
        "operationId": "ThemesController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a theme",
        "tags": [
          "Themes"
        ]
      }
    },
    "/api/workspaces/{wsId}/themes/{id}/entities": {
      "post": {
        "operationId": "ThemesController_link",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link an entity (idea/support/feature) to a theme",
        "tags": [
          "Themes"
        ]
      },
      "delete": {
        "operationId": "ThemesController_unlink",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unlink an entity from a theme",
        "tags": [
          "Themes"
        ]
      }
    },
    "/api/workspaces/{wsId}/themes/{id}/promote": {
      "post": {
        "operationId": "ThemesController_promote",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteThemeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Promote a theme to a Feature or Idea",
        "tags": [
          "Themes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pinecone API",
    "description": "Aha!-style requirements management system.\n\nI10 API 治理（2026-08-18 P2）：全局前缀 /api；认证 = Bearer JWT（Web）或 API Token（pc_tok_ 前缀，SHA-256 存储）；全局限流 100 req/min/IP（Throttler）；Swagger 仅非生产暴露。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LinkEntityDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTestPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "计划名称（如「26.3.1 回归计划」）"
          },
          "releaseId": {
            "type": "string",
            "description": "关联发布周期；空 = 跨周期专项计划"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTestPlanStatusDto": {
        "type": "object",
        "properties": {}
      },
      "AddPlanCasesDto": {
        "type": "object",
        "properties": {
          "testCaseIds": {
            "description": "用例 id 列表（与 releaseId 二选一）",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "releaseId": {
            "type": "string",
            "description": "批量拉入该发布周期的全部用例（与 testCaseIds 二选一）"
          }
        }
      },
...
      "CreateThemeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateThemeDto": {
        "type": "object",
        "properties": {}
      },
      "PromoteThemeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}